+++ ninja-core/src/main/java/ninja/Configuration.java
@@ -17,6 +17,7 @@
 package ninja;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import ninja.cache.Cache;
 import ninja.cache.CacheProvider;
 import ninja.jpa.JpaModule;
@@ -32,6 +33,7 @@ import org.slf4j.Logger;
 import com.google.inject.AbstractModule;
 import com.google.inject.Singleton;
 import ninja.utils.ObjectMapperProvider;
+import ninja.utils.XmlMapperProvider;
 
 /**
  * The basic configuration of the main ninja framework.
@@ -52,6 +54,7 @@ public class Configuration extends AbstractModule {
         System.setProperty("file.encoding", "utf-8");
         
         bind(ObjectMapper.class).toProvider(ObjectMapperProvider.class).in(Singleton.class);
+        bind(XmlMapper.class).toProvider(XmlMapperProvider.class).in(Singleton.class);
 
         bind(RouteBuilder.class).to(RouteBuilderImpl.class);