+++ ninja-core/src/main/java/ninja/RouteBuilderImpl.java
@@ -37,6 +37,7 @@ class RouteBuilderImpl implements RouteBuilder {
     private String uri;
     private Class controller;
     private Method controllerMethod;
+    private Result result;
 
     public RouteBuilderImpl GET() {
         httpMethod = "GET";
@@ -68,6 +69,11 @@ class RouteBuilderImpl implements RouteBuilder {
         this.controller = controller;
         this.controllerMethod = verifyThatControllerAndMethodExists(controller,
                 controllerMethod);
+    }
+
+    @Override
+    public void with(Result result) {
+        this.result = result;
     }
 
     @Override
@@ -153,25 +159,30 @@ class RouteBuilderImpl implements RouteBuilder {
     public Route buildRoute(Injector injector) {
         // Calculate filters
         LinkedList<Class<? extends Filter>> filters = new LinkedList<Class<? extends Filter>>();
-        filters.addAll(calculateFiltersForClass(controller));
-        FilterWith filterWith = controllerMethod
-                .getAnnotation(FilterWith.class);
-        if (filterWith != null) {
-            filters.addAll(Arrays.asList(filterWith.value()));
+        if(controller != null) {
+            filters.addAll(calculateFiltersForClass(controller));
+            FilterWith filterWith = controllerMethod
+                    .getAnnotation(FilterWith.class);
+            if (filterWith != null) {
+                filters.addAll(Arrays.asList(filterWith.value()));
+            }
         }
+
         return new Route(httpMethod, uri, controller, controllerMethod,
                 buildFilterChain(injector, filters, controller,
-                        controllerMethod));
+                        controllerMethod, result));
     }
 
     private FilterChain buildFilterChain(Injector injector,
                                          LinkedList<Class<? extends Filter>> filters,
                                          Class<?> controller,
-                                         Method controllerMethod) {
+                                         Method controllerMethod,
+                                         Result result) {
 
         if (filters.isEmpty()) {
 
-            return new FilterChainEnd(injector.getProvider(controller),
+            return result != null ? new FilterChainEnd(result) :
+                    new FilterChainEnd(injector.getProvider(controller),
                     ControllerMethodInvoker.build(controllerMethod, injector));
 
         } else {
@@ -180,7 +191,7 @@ class RouteBuilderImpl implements RouteBuilder {
 
             return new FilterChainImpl(injector.getProvider(filter),
                     buildFilterChain(injector, filters, controller,
-                            controllerMethod));
+                            controllerMethod, result));
 
         }
     }